boto_test/pyproject.toml
[project]
name = "boto-shortener"
version = "0.1.0"
description = "Тестовое задание: Boto"
requires-python = ">=3.12"
dependencies = [
    "fastapi",
    "uvicorn",
    "pydantic-settings",
    "python-dotenv",
]

[project.optional-dependencies]
test = [
    "pytest",
    "httpx",  # Нужен для тестирования FastAPI через TestClient
]

[tool.pytest.ini_options]
testpaths = ["tests"]
python_files = "test_*.py"
